requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
astor==0.8.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
et-xmlfile==1.1.0
execnet==1.9.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==6.2.5
pytest-ast-transformer==1.0.3
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-soft-assertions==0.1.2
pytest-xdist==2.5.0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
texttable==1.6.4
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0
